Analyze more tests using PHPStan (#20872)
