+++ phpstan.dist.neon
@@ -13,25 +13,14 @@ parameters:
     excludePaths:
         analyse:
             - tests/data
-            - tests/framework/base/stub
-            - tests/framework/console/stubs
-            - tests/framework/di/stubs
-            - tests/framework/filters/stubs
-            - tests/framework/web/mocks
-            - tests/framework/web/stubs
+            - tests/framework/*/stubs/*
+            - tests/framework/*/mocks/*
             # TODO: Analyze these directories
             - tests/framework/db
             - tests/framework/grid
             - tests/framework/helpers
             - tests/framework/i18n
             - tests/framework/log
-            - tests/framework/mail
-            - tests/framework/models
-            - tests/framework/mutex
-            - tests/framework/rbac
-            - tests/framework/requirements
-            - tests/framework/rest
-            - tests/framework/test
             - tests/framework/validators
 
     bootstrapFiles:
