+++ tests/framework/base/ComponentTest.php
@@ -507,7 +507,7 @@ class ComponentTest extends TestCase
      */
     public function testEventClosureDetachesItself(): void
     {
-        $obj = require __DIR__ . '/stub/AnonymousComponentClass.php';
+        $obj = require __DIR__ . '/stubs/AnonymousComponentClass.php';
 
         $obj->trigger('barEventOnce');
         $this->assertEquals(1, $obj->foo);
