+++ tests/framework/base/ModelTest.php
@@ -508,7 +508,7 @@ class ModelTest extends TestCase
 
     public function testFormNameWithAnonymousClass(): void
     {
-        $model = require __DIR__ . '/stub/AnonymousModelClass.php';
+        $model = require __DIR__ . '/stubs/AnonymousModelClass.php';
 
         $this->expectException('yii\base\InvalidConfigException');
         $this->expectExceptionMessage('The "formName()" method should be explicitly defined for anonymous models');
