+++ tests/framework/base/stubs/AnonymousComponentClass.php
