+++ tests/framework/base/stubs/AnonymousModelClass.php
