+++ tests/framework/rbac/DbManagerTestCase.php
@@ -30,6 +30,7 @@ use yiiunit\framework\log\ArrayTarget;
  * DbManagerTestCase.
  * @group db
  * @group rbac
+ * @extends ManagerTestCase<DbManager>
  */
 abstract class DbManagerTestCase extends ManagerTestCase
 {
@@ -37,7 +38,7 @@ abstract class DbManagerTestCase extends ManagerTestCase
     protected static $driverName;
 
     /**
-     * @var Connection
+     * @var Connection|null
      */
     protected $db;
 
@@ -50,7 +51,7 @@ abstract class DbManagerTestCase extends ManagerTestCase
         $this->auth = $this->createManager();
 
         $this->assertEquals([], $this->auth->getUserIdsByRole('nonexisting'));
-        $this->assertEquals(['123', 'reader A'], $this->auth->getUserIdsByRole('reader'), '', 0.0, 10, true);
+        $this->assertEquals(['123', 'reader A'], $this->auth->getUserIdsByRole('reader'));
         $this->assertEquals(['author B'], $this->auth->getUserIdsByRole('author'));
         $this->assertEquals(['admin C'], $this->auth->getUserIdsByRole('admin'));
     }
@@ -158,9 +159,6 @@ abstract class DbManagerTestCase extends ManagerTestCase
         return $db;
     }
 
-    /**
-     * @return ManagerInterface
-     */
     protected function createManager()
     {
         return new DbManager(['db' => $this->getConnection(), 'defaultRoles' => ['myDefaultRole']]);
