+++ tests/framework/rbac/PhpManagerTest.php
@@ -33,7 +33,7 @@ use Yii;
 
 /**
  * @group rbac
- * @property ExposedPhpManager $auth
+ * @extends ManagerTestCase<ExposedPhpManager>
  */
 class PhpManagerTest extends ManagerTestCase
 {
